+++ layrz_sdk/entities/sensor.py
@@ -9,3 +9,7 @@ class Sensor(BaseModel):
   pk: int = Field(description='Defines the primary key of the sensor')
   name: str = Field(description='Defines the name of the sensor')
   slug: str = Field(description='Defines the slug of the sensor')
+  formula: str = Field(
+    default='',
+    description='Defines the formula of the sensor, used for calculations',
+  )