+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.1.49
 
-- New models `AtsExitExecutionHistory` and  `AtsPossibleExit` for Ats exits.
+- New models `AtsExitExecutionHistory` and  `AtsPossibleExit` for Ats exits add formula field to `Sensor`.
 ## 3.1.48
 
 - Moved `DestinationPhone` entity to its own file, and renamed from `TwilioHostPhone` to `DestinationPhone`.